% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selfea.R
\name{calculate_cohen_f2}
\alias{calculate_cohen_f2}
\title{calculate_cohen_f2}
\usage{
calculate_cohen_f2(model_glm, df_aov)
}
\arguments{
\item{model_glm}{GLM model generated by 'glm' function}

\item{df_aov}{A data frame containing groups in 'Run' column and values in 'SC' column}
}
\value{
Cohen's f2 (an effect size for linear models)
}
\description{
Calculate Cohen's f2. Followed formulars at wikipages (https://en.wikipedia.org/wiki/Effect_size , https://en.wikipedia.org/wiki/Coefficient_of_determination)
}

