% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selfea.R
\name{draw_scatter_plots}
\alias{draw_scatter_plots}
\title{draw_scatter_plots}
\usage{
draw_scatter_plots(input_data_frame, max_pvalue, min_ES, power_desired, x_label,
  y_label)
}
\arguments{
\item{input_data_frame}{A data frame that consists of 'x' (P-value), 'y' (Effect size),'cat' (significant or not).}

\item{max_pvalue}{P-value threshold}

\item{min_ES}{Effect size filter threshold}

\item{power_desired}{Give the statistical power you desired for output significant list}

\item{x_label}{Label of X axis}

\item{y_label}{Label of y axis}
}
\value{
A scatter plot
}
\description{
Draw a scatterplot to show how significant IDs are distinguished from the total
}

