% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selfea.R
\name{glm_anova}
\alias{glm_anova}
\title{glm_anova}
\usage{
glm_anova(dataset.expr, dataset.ID, group, padj = "fdr")
}
\arguments{
\item{dataset.expr}{A data frame that has column names for distinguishing experiments and numerical values for expression levels}

\item{dataset.ID}{A vector of the obtained expression profile's ID column}

\item{group}{A data frame that consists of 'Col_Name' and 'Group' obtained from the user file through get_statistics_from_file.}

\item{padj}{Choose one of these c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none").
"fdr" is default option.}
}
\value{
A data frame containing ID, Cohen's W, Cohen's F, Max fold change,
   GLM Negative Binomial P-value, GLM Quasi-Poisson P-value and ANOVA with Normal P-value.
}
\description{
Calculate P-values from ANOVA using Normal, Quasi-Poisson and Negative Binomial distribution and Cohen's effect sizes
}

